+++ config/environment.d.ts
@@ -154,6 +154,9 @@ declare const config: {
         navigation: {
             institutions: string;
         };
+        ABTesting: {
+            homePageVersionB: string;
+        };
         storageI18n: string;
         enableInactiveSchemas: string;
         verifyEmailModals: string;