+++ CHANGELOG.md
@@ -5,6 +5,18 @@ The format is based on [Keep a Changelog](http://keepachangelog.com/en/1.0.0/)
 and this project adheres to [Semantic Versioning](http://semver.org/spec/v2.0.0.html).
 
 ## [Unreleased]
+### Added
+- Components:
+    - `get-started-button` - a button that takes you to the '/register' page.
+    - `search-bar` - a search bar component that takes you to the search page.
+    - `new-home/-components/hero-banner` - a banner to be used on the logged-out homepage.
+- Tests:
+    - Acceptance:
+        - `new-home`
+    - Integration:
+        - `get-started-button`
+        - `search-bar`
+        - `hero-banner`
 
 ## [19.4.0] - 2019-04-25
 ### Added